move data structure definition
