+++ Data Structures/Stack/stack.js
